% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{add_call_callback}
\alias{add_call_callback}
\title{Add a function-specific callback to the call callbacks.}
\usage{
add_call_callback(callbacks, fn, cb)
}
\arguments{
\item{callbacks}{The existing callbacks.}

\item{fn}{The function to which calls should be modified.}

\item{cb}{The callback function to invoke.}
}
\value{
\preformatted{     The updated callbacks.
}
}
\description{
This function adds to the existing call callback, rather than replace it,
by putting a callback in front of it to be tested first. The callback will
be invoked when the traversal sees a call to a specific function.
}
