% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static-analysis.R
\docType{data}
\name{annotate_assigned_symbols_callbacks}
\alias{annotate_assigned_symbols_callbacks}
\title{Put attribute "assigned_symbols" on expressions bottom-up}
\format{An object of class \code{list} of length 7.}
\usage{
annotate_assigned_symbols_callbacks
}
\description{
Put attribute "assigned_symbols" on expressions bottom-up
}
\keyword{datasets}
