% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth-first-traversals.R
\name{depth_first_analyse_expr}
\alias{depth_first_analyse_expr}
\title{Analyse an expression.}
\usage{
depth_first_analyse_expr(expr, callbacks, topdown, ...)
}
\arguments{
\item{expr}{An R expression}

\item{callbacks}{List of callbacks to apply.}

\item{topdown}{A list of additional information gathered in the traversal.}

\item{...}{Additional data that will be passed along to callbacks.}
}
\value{
The result of the last bottom-up traversal.
}
\description{
Traverses the expression \code{expr} depth-first and analyse it it using \code{callbacks}.
}
\seealso{
analysis_callbacks

identity_analysis_callback

depth_first_analyse_function
}
