% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth-first-traversals.R
\name{depth_first_analyse_function}
\alias{depth_first_analyse_function}
\title{Analyse the body of function.}
\usage{
depth_first_analyse_function(fn, callbacks, topdown = list(), ...)
}
\arguments{
\item{fn}{A (closure) function.}

\item{callbacks}{List of callbacks to apply.}

\item{topdown}{A list of additional information that will be considered
top-down in the traversal.}

\item{...}{Additional data that will be passed along to callbacks.}
}
\value{
The result of the last bottom-up call to a callback.
}
\description{
Traverses the body of \code{fn} and analyse it based on \code{callbacks}.
}
\seealso{
depth_first_analyse_expr

depth_first_rewrite_function

analysis_callbacks
}
