% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth-first-traversals.R
\name{depth_first_rewrite_function}
\alias{depth_first_rewrite_function}
\title{Transform the body of function.}
\usage{
depth_first_rewrite_function(fn, callbacks, topdown = list(),
  wflags = warning_flags(), ...)
}
\arguments{
\item{fn}{A (closure) function.}

\item{callbacks}{List of callbacks to apply.}

\item{topdown}{A list of additional information that will be considered
top-down in the traversal.}

\item{wflags}{Warning flags, see \code{\link[=warning_flags]{warning_flags()}}.}

\item{...}{Additional data that will be passed along to callbacks.}
}
\value{
A new function similar to \code{fn} but with a transformed body.
}
\description{
Traverses the body of \code{fn} and rewrite it based on \code{callbacks}.
}
\seealso{
depth_first_rewrite_expr

rewrite_callbacks
}
