% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{identity_callback}
\alias{identity_callback}
\title{A callback that does not do any transformation.}
\usage{
identity_callback(expr, env, params)
}
\arguments{
\item{expr}{The expression to (not) transform.}

\item{env}{The environment of the function the expression is inside.}

\item{params}{The formal parameters of the function.}
}
\value{
\code{expr}
}
\description{
A callback that does not do any transformation.
}
