% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-dsl.R
\name{make_transform_function}
\alias{make_transform_function}
\title{Create a transformation function from a callback specification.}
\usage{
make_transform_function(callbacks)
}
\arguments{
\item{callbacks}{The callbacks to apply to transformed functions.}
}
\value{
A function that can transform other functions.
}
\description{
Create a transformation function from a callback specification.
}
