% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{make_with_callback}
\alias{make_with_callback}
\title{Create a function for setting callbacks.}
\usage{
make_with_callback(cb_name)
}
\arguments{
\item{cb_name}{The name of the callback to set}
}
\value{
A function that can be used in a pipe to set a callback.
}
\description{
Create a function for setting callbacks.
}
