% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{merge_bottomup}
\alias{merge_bottomup}
\title{Merge the results of several bottomup results.}
\usage{
merge_bottomup(bottomup)
}
\arguments{
\item{bottomup}{List of bottom up analysis results.}
}
\description{
The \code{bottomup} parameter in callbacks will be calculated for all parameters
of \code{call`` expressions. The parameter to the callbacks are thus a list of lists. This function merges these lists into one that contain a list for each named component in the }bottomup\code{parameter. If results are not named in the}bottomup` list, they are discarded.
}
\details{
The vectors from \code{bottomup} are concatenated, so one level of lists
will be flattened. Use more lists, like \code{list(list(2), list(3))}
if you want to prevent this.
}
\seealso{
depth_first_analyse_function

depth_first_analyse_expr
}
