% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formals.R
\name{remove_formal_}
\alias{remove_formal_}
\title{Remove a parameter from the formal parameters of a function.}
\usage{
remove_formal_(fn, par)
}
\arguments{
\item{fn}{A function we are modifying}

\item{par}{A parameter of \code{fn} (should be in \code{formals(fn)} and be quoted)}
}
\value{
A modified function equal to \code{fn} but with \code{par} removed from the formal
parameters.
}
\description{
Remove a parameter from the formal parameters of a function.
}
