% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{rewrite_callbacks}
\alias{rewrite_callbacks}
\alias{analysis_callbacks}
\alias{with_atomic_callback}
\alias{with_pairlist_callback}
\alias{with_symbol_callback}
\alias{with_primitive_callback}
\alias{with_call_callback}
\alias{with_topdown_callback}
\title{Default expression-transformation callbacks.}
\usage{
rewrite_callbacks()

analysis_callbacks()

with_atomic_callback(callbacks, fn)

with_pairlist_callback(callbacks, fn)

with_symbol_callback(callbacks, fn)

with_primitive_callback(callbacks, fn)

with_call_callback(callbacks, fn)

with_topdown_callback(callbacks, fn)
}
\arguments{
\item{callbacks}{The list of callbacks}

\item{fn}{A function to install as a callback.}
}
\description{
Callbacks must be functions that take three arguments: The expression
to rewrite, the environment of the function we are rewriting (i.e. the
environment it is defined in, not the function call frame), and a list of
formal parameters of the function we are translating.
}
\section{Functions}{
\itemize{
\item \code{rewrite_callbacks}: Default callbacks for rewriting expressions

\item \code{analysis_callbacks}: Default callbacks for analysing expressions

\item \code{with_atomic_callback}: Set the atomic callback function.

\item \code{with_pairlist_callback}: Set the pairlist callback function.

\item \code{with_symbol_callback}: Set the symbol callback function.

\item \code{with_primitive_callback}: Set the primitive callback function.

\item \code{with_call_callback}: Set the call callback function.

\item \code{with_topdown_callback}: Set the topdown information passing callback
function.
}}

\seealso{
with_atomic_callback

with_pairlist_callback

with_symbol_callback

with_primitive_callback

with_call_callback

with_topdown_callback
}
