% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition-handling.R
\name{warning_flags}
\alias{warning_flags}
\alias{set_warn_on_unknown_function}
\alias{unset_warn_on_unknown_function}
\alias{set_warn_on_local_function}
\alias{unset_warn_on_local_function}
\title{Collection of warning flags used when traversing expressions.}
\usage{
warning_flags()

set_warn_on_unknown_function(flags)

unset_warn_on_unknown_function(flags)

set_warn_on_local_function(flags)

unset_warn_on_local_function(flags)
}
\arguments{
\item{flags}{Used when setting or unsetting flags.}
}
\description{
Collection of warning flags used when traversing expressions.
}
\section{Functions}{
\itemize{
\item \code{set_warn_on_unknown_function}: Enable warnings when encountring an unknown
function

\item \code{unset_warn_on_unknown_function}: Disable warnings when encountring an unknown
function

\item \code{set_warn_on_local_function}: Enable warnings when encountring a local variable
with a name that matches one installed for transformation.

\item \code{unset_warn_on_local_function}: Disable warnings when encountring a local variable
with a name that matches one installed for transformation.
}}

