% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static-analysis.R
\docType{data}
\name{annotate_bound_variables_callbacks}
\alias{annotate_bound_variables_callbacks}
\title{Propagate parameters and local variables top-down to assign attribute
"bound" to all call expressions.}
\format{An object of class \code{list} of length 7.}
\usage{
annotate_bound_variables_callbacks
}
\description{
Propagate parameters and local variables top-down to assign attribute
"bound" to all call expressions.
}
\keyword{datasets}
