% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\docType{data}
\name{callbacks}
\alias{callbacks}
\alias{with_atomic_callback}
\alias{with_pairlist_callback}
\alias{with_symbol_callback}
\alias{with_primitive_callback}
\alias{with_call_callback}
\title{Default expression-transformation callbacks.}
\format{An object of class \code{list} of length 5.}
\usage{
callbacks

with_atomic_callback(callbacks, fn)

with_pairlist_callback(callbacks, fn)

with_symbol_callback(callbacks, fn)

with_primitive_callback(callbacks, fn)

with_call_callback(callbacks, fn)
}
\description{
Default expression-transformation callbacks.
}
\section{Functions}{
\itemize{
\item \code{with_atomic_callback}: Set the atomic callback function.

\item \code{with_pairlist_callback}: Set the pairlist callback function.

\item \code{with_symbol_callback}: Set the symbol callback function.

\item \code{with_primitive_callback}: Set the primitive callback function.

\item \code{with_call_callback}: Set the call callback function.
}}

\seealso{
with_atomic_callback

with_pairlist_callback

with_symbol_callback

with_primitive_callback

with_call_callback
 @export
}
\keyword{datasets}
