% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth-first-transformation.R
\name{depth_first_rewrite_expr}
\alias{depth_first_rewrite_expr}
\title{Transform an expression.}
\usage{
depth_first_rewrite_expr(expr, callbacks, topdown, ...)
}
\arguments{
\item{expr}{An R expression}

\item{callbacks}{List of callbacks to apply.}

\item{topdown}{A list of additional information gathered in the traversal.}

\item{...}{Additional data that will be passed along to callbacks.}
}
\value{
A modified expression.
}
\description{
Traverses the expression \code{expr} depth-first and transform it using \code{callbacks}.
}
\seealso{
callbacks

identity_callback

depth_first_rewrite_function
}
