% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth-first-transformation.R
\name{depth_first_rewrite_expr}
\alias{depth_first_rewrite_expr}
\title{Transform an expression.}
\usage{
depth_first_rewrite_expr(expr, callbacks, env, param)
}
\arguments{
\item{expr}{An R expression}

\item{callbacks}{List of callbacks to apply.}

\item{fun_env}{The environment of the function
the expression will be evaluated in.}

\item{params}{The formal parameters of the function
the expression is evaluated in.}
}
\value{
A modified expression.
}
\description{
Traverses the expression \code{expr} depth-first and transform it using \code{callbacks}.
}
\seealso{
callbacks

depth_first_rewrite_function
}
