% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth-first-traversals.R
\name{depth_first_rewrite_function}
\alias{depth_first_rewrite_function}
\title{Transform the body of function.}
\usage{
depth_first_rewrite_function(fn, callbacks, topdown = list(), ...)
}
\arguments{
\item{fn}{A (closure) function.}

\item{callbacks}{List of callbacks to apply.}

\item{topdown}{A list of additional information that will be considerd
top-down in the traversal.}

\item{...}{Additional data that will be passed along to callbacks.}
}
\value{
A new function similar to \code{fn} but with a transformed body.
}
\description{
Traverses the body of \code{fn} and rewrite it based on \code{callbacks}.
}
\examples{
log_calls_transformation <- function(call_expr, env, param, ...) {
    call_fn <- eval(call_expr[[1]], envir = env)
    if (rlang::is_primitive(call_fn)) return(call_expr)
    if (!rlang::is_symbol(call_expr[[1]])) return(call_expr)

    rlang::expr({
        cat("Calling", rlang::UQ(as.character(call_expr[[1]])), "\\n")
        rlang::UQ(call_expr)
    })
}

f <- function(x) {
    if (x > 0) f(x - 1)
    else print("Done")
}

cb <- rewrite_callbacks() \%>\% with_call_callback(log_calls_transformation)
f <- depth_first_rewrite_function(f, cb)

}
\seealso{
depth_first_rewrite_expr

rewrite_callbacks
}
