% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{identity_callback}
\alias{identity_callback}
\title{A callback that does not do any transformation.}
\usage{
identity_callback(expr, ...)
}
\arguments{
\item{expr}{The expression to (not) transform.}

\item{...}{Additional named parameters.}
}
\value{
\code{expr}
}
\description{
Callbacks have one required argument, \code{expr}, but will actually
be called with more. The additional named parameters are:
\itemize{
\item \strong{env}      The function environment of the function we are transforming
\item \strong{params}   The formal parameters of the function we are transforming
\item \strong{topdown}  Data passed top-down in the traversal.
\item \strong{buttomup} Data collected by depth-first traversals before a callback
is called.
plus whatever the user provide to \code{\link[=depth_first_rewrite_function]{depth_first_rewrite_function()}}.
}
}
