% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{identity_callback}
\alias{identity_callback}
\title{A callback that does not do any transformation.}
\usage{
identity_callback(expr, env, param)
}
\arguments{
\item{expr}{The expression to (not) transform.}

\item{param}{The formal parameters of the function.}

\item{evn}{The environment of the function the expression is inside.}
}
\value{
\code{expr}
}
\description{
A callback that does not do any transformation.
}
