% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-dsl.R
\name{rewrite_with}
\alias{rewrite_with}
\alias{rewrite}
\title{Functions for applying a sequence of rewrites.}
\usage{
rewrite_with(fn, callbacks, ...)

rewrite(fn)
}
\arguments{
\item{fn}{The function to rewrite}

\item{callbacks}{The callbacks that should do the rewriting}

\item{...}{Additional parameters passed along to the callbacks.}
}
\value{
A rewritten function
}
\description{
The \code{\link[=rewrite]{rewrite()}} function applies a series of transformations to an input
function, \code{fn} and returns the result. This result can then be used in a
pipeline of \code{\link[=rewrite_with]{rewrite_with()}} calls for further analysis.
}
\details{
The flow of transformations goes starts with \code{\link[=rewrite]{rewrite()}} and is followed
by a series of \code{\link[=rewrite_with]{rewrite_with()}} for additional rewrite callbacks.
}
\section{Functions}{
\itemize{
\item \code{rewrite_with}: Apply \code{callbacks} over \code{fn} to rewrite it.

\item \code{rewrite}: Function for starting a rewrite.
}}

\examples{
f <- function(x) 2 + x
cb <- rewrite_callbacks() \%>\%
   add_call_callback(f, function(expr, ...) {
       quote(2 + x)
   })
tr_f <- . \%>\% rewrite() \%>\% rewrite_with(cb)

g <- function(y) y + f(y) # body(g) is now quote(y + (2 + x))

}
\seealso{
rewrite_callbacks
}
