% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-dsl.R
\docType{data}
\name{rewrites}
\alias{rewrites}
\title{Object for setting up a transformation pipeline when defining functions}
\format{An object of class \code{foolbox_rewrite_spec} of length 1.}
\usage{
rewrites
}
\description{
Object for setting up a transformation pipeline when defining functions
}
\keyword{datasets}
