% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-dsl.R
\name{[.foolbox_rewrite_spec}
\alias{[.foolbox_rewrite_spec}
\title{Provide list of rewrite transformations.}
\usage{
\method{[}{foolbox_rewrite_spec}(dummy, ...)
}
\arguments{
\item{dummy}{The dummy-table \code{\link{rewrites}}. It is only here because it
allows us to use subscripts as part of the domain-specific language.}

\item{...}{A list of rewrite functions.}
}
\description{
This subscript operator is used together with \code{\link{rewrites}} to specify
a sequence of transformations to apply to a new function we define.
}
